Split the ABIs into separate implementations and trait files.
