+++ lib/compiler-llvm/src/lib.rs
@@ -14,6 +14,7 @@
 #![doc(html_favicon_url = "https://wasmer.io/static/icons/favicon.ico")]
 #![doc(html_logo_url = "https://avatars3.githubusercontent.com/u/44205449?s=200&v=4")]
 
+mod abi;
 mod compiler;
 mod config;
 mod object_file;
