+++ lib/compiler-llvm/src/translator/mod.rs
@@ -1,4 +1,3 @@
-pub mod abi;
 mod code;
 pub mod intrinsics;
 //mod stackmap;
